refactor(cli): split diagnostics_cli into a diagnostics package
The 862-line diagnostics_cli.py mixed five concerns in one module. Split
into a cohesive subpackage:

- diagnostics/models.py     — Finding / DiagnosticReport records
- diagnostics/loaders.py    — dotenv bootstrap + shared JSON loader
- diagnostics/collectors.py — the per-section _collect_* probes
- diagnostics/render.py      — human-readable console rendering
- diagnostics/commands.py    — handle_doctor / config-validate entry points

diagnostics_cli.py is now a thin (46-line) back-compat shim re-exporting the
public handlers plus the records/collectors that agentao.cli and the test
suite import by name, so the lazy import in cli/__init__.py and existing
`from agentao.cli.diagnostics_cli import ...` call sites keep working with
zero changes. Behavior-preserving; full suite green (2840 passed).
